GlucoRisk_Package/requirements.txt
pyserial
rich
flask
werkzeug
numpy
scikit-learn
twilio
python-dotenv
flask-wtf
flask-limiter
reportlab
gunicorn
paho-mqtt
cryptography